requirements.txt
# Snowflake Semantic Model Pipeline Requirements
# Install with: pip install -r requirements.txt

# Core Snowflake connector
snowflake-connector-python>=3.0.0

# Data processing
pandas>=1.3.0
numpy>=1.21.0

# YAML processing
PyYAML>=6.0

# HTTP requests for GitHub API
requests>=2.25.0

# Logging and utilities
python-dateutil>=2.8.0
pytz>=2021.1

# Optional: For better performance with large datasets
# pyarrow<19.0.0  # Uncomment if you need Arrow support

# Development and testing (optional)
# pytest>=6.0.0
# black>=21.0.0
# flake8>=3.9.0